reinforcement/minigo/tensorflow2/minigo/requirements.txt
absl-py==0.9.0
autopep8==1.5.3
choix==0.3.3
fire==0.3.1
google-cloud-logging==1.15.0
google-cloud-bigtable==1.2.1
grpcio-tools==1.29.0
keras
numpy
protobuf==3.12.2
pylint==2.5.3
sgf==0.5
six==1.13.0
tqdm==4.46.1
pyasn1==0.4.8
setuptools==47.1.1

oauth2client==4.1

mpi4py
